# services/wallet-api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Giano Wallet API",
    "description": "WebAuthn ceremonies, sessions and policied ERC-4337 user-operation relay for Giano smart wallets (multi-tenant).",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque session token from a ceremony verify endpoint"
      },
      "adminKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-tenant admin API key (provisioned via TENANTS_SEED)"
      }
    },
...
  },
  "paths": {
...
    "/v1/userops": {
      "post": {
        "tags": [
          "userops"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userOperation": {
                    "type": "object",
                    "properties": {
                      "sender": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]{40}$"
                      },
                      "nonce": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]+$"
                      },
                      "callData": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]*$"
                      },
                      "callGasLimit": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]+$"
                      },
                      "verificationGasLimit": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]+$"
                      },
                      "preVerificationGas": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]+$"
                      },
                      "maxFeePerGas": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]+$"
                      },
                      "maxPriorityFeePerGas": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]+$"
                      },
                      "signature": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]*$"
                      },
                      "factory": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]{40}$"
                      },
                      "factoryData": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]*$"
                      },
                      "paymaster": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]{40}$"
                      },
                      "paymasterVerificationGasLimit": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]+$"
                      },
                      "paymasterPostOpGasLimit": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]+$"
                      },
                      "paymasterData": {
                        "type": "string",
                        "pattern": "^0x[0-9a-fA-F]*$"
                      }
                    },
                    "required": [
                      "sender",
                      "nonce",
                      "callData",
                      "callGasLimit",
                      "verificationGasLimit",
                      "preVerificationGas",
                      "maxFeePerGas",
                      "maxPriorityFeePerGas",
                      "signature"
                    ],
                    "additionalProperties": false
                  }
                },
                "required": [
                  "userOperation"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userOperationHash": {
                      "type": "string"
                    },
                    "duplicate": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "userOperationHash"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "policy": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "rule": {
                            "type": "string"
                          },
                          "passed": {
                            "type": "boolean"
                          },
                          "detail": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "rule",
                          "passed"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message",
                    "policy"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/userops/{hash}": {
      "get": {
        "tags": [
          "userops"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-fA-F]{64}$"
            },
            "in": "path",
            "name": "hash",
            "required": true
          }
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userOperationHash": {
                      "type": "string"
                    },
                    "sender": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "accepted",
                        "rejected",
                        "submitted",
                        "failed"
                      ]
                    },
                    "policyResults": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "rule": {
                            "type": "string"
                          },
                          "passed": {
                            "type": "boolean"
                          },
                          "detail": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "rule",
                          "passed"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "rejectReason": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "userOperationHash",
                    "sender",
                    "status",
                    "policyResults",
                    "rejectReason",
                    "createdAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/userops/{hash}/receipt": {
      "get": {
        "tags": [
          "userops"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-fA-F]{64}$"
            },
            "in": "path",
            "name": "hash",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "receipt": {
                      "nullable": true
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  }
}